cloud_functions/train-minute-yield-curve/requirements.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.11.0
DateTime==5.5
db-dtypes==1.2.0
Deprecated==1.2.15
deprecation==2.1.0
finnhub-python==2.4.24
Flask==3.0.3
functions-framework==3.8.0
google-api-core==2.19.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-bigquery==3.24.0
google-cloud-bigquery-storage==2.26.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-cloud-secret-manager==2.20.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.14.0
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
idna==3.7
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.29.0
packaging==24.0
pandas==1.5.3
pandas-gbq==0.23.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydata-google-auth==1.8.2
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.0.0
six==1.16.0
tabulate==0.9.0
threadpoolctl==3.5.0
urllib3==2.2.1
watchdog==4.0.1
Werkzeug==3.0.3
wrapt==1.17.0
xlrd==2.0.1
zipp==3.21.0
zope.interface==6.4.post2